+++ myems-admin/app/translations.js
@@ -309,7 +309,11 @@ function config($translateProvider) {
                 DATA_SOURCE_LIST: 'Data Source List',
 
                 CONTROL_MODE_LIST: 'Control Mode List',
-                END_TIME_SHOULD_BE_AFTER_START_TIME: 'The end time must be after the start time',
+
+				END_TIME_SHOULD_BE_AFTER_START_TIME: 'The end time must be after the start time',
+				TARIFF_TIME_PERIODS_OVERLAP: 'Only one pricing method is allowed in the same time period',
+				VALID_THROUGH_TIME_SHOULD_BE_AFTER_VALID_FROM_TIME: 'The end date of the rate validity period must be after the start date',
+
                 IS_ACTIVE: 'Is Active',
                 START_TIME: 'Start Time',
                 END_TIME: 'End Time',
@@ -1166,6 +1170,7 @@ function config($translateProvider) {
             METER: {
                 HOURLY_LOW_LIMIT: 'Hourly Low Limit (Inclusive)',
                 INPUT_HOURLY_LOW_LIMIT: 'Please Input Hourly Low Limit',
+                INVALID_INPUT_HINT: 'The entered value is incorrect, with a range of 0 to 999999',
                 HOURLY_HIGH_LIMIT: 'Hourly High Limit (Inclusive)',
                 INPUT_HOURLY_HIGH_LIMIT: 'Please Input Hourly High Limit',
                 MASTER_METER: 'Master Meter',
@@ -2264,7 +2269,11 @@ function config($translateProvider) {
                 DATA_SOURCE_LIST: '数据源列表',
 
                 CONTROL_MODE_LIST: '控制模式列表',
-                END_TIME_SHOULD_BE_AFTER_START_TIME: '结束时间应在开始时间之后',
+
+				END_TIME_SHOULD_BE_AFTER_START_TIME: '结束时间应在开始时间之后',
+				TARIFF_TIME_PERIODS_OVERLAP:'同一时段只允许有一种计价',
+				VALID_THROUGH_TIME_SHOULD_BE_AFTER_VALID_FROM_TIME: '费率有效期结束期应在开始期之后',
+
                 IS_ACTIVE: '是否开启',
                 START_TIME: '开始时间',
                 END_TIME: '结束时间',
@@ -3117,6 +3126,7 @@ function config($translateProvider) {
             METER: {
                 HOURLY_LOW_LIMIT: '每小时最小值（包含）',
                 INPUT_HOURLY_LOW_LIMIT: '请输入每小时最小值',
+                INVALID_INPUT_HINT: '输入的数值不正确，数值范围为0到999999',
                 HOURLY_HIGH_LIMIT: '每小时最大值（包含）',
                 INPUT_HOURLY_HIGH_LIMIT: '请输入每小时最大值',
                 MASTER_METER: '上级计量表',
@@ -4218,7 +4228,11 @@ function config($translateProvider) {
                 DATA_SOURCE_LIST: 'Datenquellenliste',
 
                 CONTROL_MODE_LIST: 'Control Mode List',
-                END_TIME_SHOULD_BE_AFTER_START_TIME: 'Die Endzeit muss nach der Startzeit liegen',
+
+				END_TIME_SHOULD_BE_AFTER_START_TIME: 'Die Endzeit muss nach der Startzeit liegen',
+				TARIFF_TIME_PERIODS_OVERLAP:'Im selben Zeitraum ist nur ein Preisberechnungssystem erlaubt',
+				VALID_THROUGH_TIME_SHOULD_BE_AFTER_VALID_FROM_TIME: 'Das Ende der Tarifgültigkeit muss nach dem Beginn liegen',
+
                 IS_ACTIVE: 'Is Active',
                 START_TIME: 'Start Time',
                 END_TIME: 'End Time',
@@ -5073,6 +5087,7 @@ function config($translateProvider) {
             METER: {
                 HOURLY_LOW_LIMIT: 'Stündliches Minimum (inklusive)',
                 INPUT_HOURLY_LOW_LIMIT: 'Bitte geben Sie das stündliche Minimum ein',
+                INVALID_INPUT_HINT: 'Der eingegebene Wert ist falsch und liegt im Bereich von 0 bis 999999.',
                 HOURLY_HIGH_LIMIT: 'Stundenmaximum (inklusive)',
                 INPUT_HOURLY_HIGH_LIMIT: 'Bitte geben Sie das Stundenmaximum ein',
                 MASTER_METER: 'Überlegenes Messgerät',
@@ -6166,7 +6181,11 @@ function config($translateProvider) {
                 DATA_SOURCE_LIST: 'Liste des sources de données',
 
                 CONTROL_MODE_LIST: 'Control Mode List',
-                END_TIME_SHOULD_BE_AFTER_START_TIME: 'L’heure de fin doit être après l’heure de début',
+
+				END_TIME_SHOULD_BE_AFTER_START_TIME: 'L’heure de fin doit être après l’heure de début',
+				TARIFF_TIME_PERIODS_OVERLAP:'Un seul mode de tarification est autorisé pour la même période',
+				VALID_THROUGH_TIME_SHOULD_BE_AFTER_VALID_FROM_TIME: 'La date de fin de validité du tarif doit être postérieure à la date de début',
+
                 IS_ACTIVE: 'Is Active',
                 START_TIME: 'Start Time',
                 END_TIME: 'End Time',
@@ -7022,6 +7041,7 @@ function config($translateProvider) {
             METER: {
                 HOURLY_LOW_LIMIT: 'Limite horaire basse (inclus)',
                 INPUT_HOURLY_LOW_LIMIT: 'Veuillez saisir la limite horaire basse',
+                INVALID_INPUT_HINT: 'Les valeurs entrées sont incorrectes, les valeurs vont de 0 à 999999',
                 HOURLY_HIGH_LIMIT: 'Limite horaire élevée (intégratrice)',
                 INPUT_HOURLY_HIGH_LIMIT: 'Veuillez saisir la limite horaire élevée',
                 MASTER_METER: 'Compteur principal',
@@ -8118,7 +8138,11 @@ function config($translateProvider) {
                 DATA_SOURCE_LIST: 'Lista de orígenes de datos',
 
                 CONTROL_MODE_LIST: 'Control Mode List',
-                END_TIME_SHOULD_BE_AFTER_START_TIME: 'La hora de finalización debe ser posterior a la hora de inicio',
+
+				END_TIME_SHOULD_BE_AFTER_START_TIME: 'La hora de finalización debe ser posterior a la hora de inicio',
+				TARIFF_TIME_PERIODS_OVERLAP:'Solo se permite un método de tarificación en el mismo período',
+				VALID_THROUGH_TIME_SHOULD_BE_AFTER_VALID_FROM_TIME: 'La fecha de fin de vigencia de la tarifa debe ser posterior a la fecha de inicio',
+
                 IS_ACTIVE: 'Is Active',
                 START_TIME: 'Start Time',
                 END_TIME: 'End Time',
@@ -8974,6 +8998,7 @@ function config($translateProvider) {
             METER: {
                 HOURLY_LOW_LIMIT: 'Límite bajo por hora (Contener)',
                 INPUT_HOURLY_LOW_LIMIT: 'Por favor, introduzca el límite bajo por hora',
+                INVALID_INPUT_HINT: 'Los valores introducidos no son correctos, el rango de valores es de 0 a 999999',
                 HOURLY_HIGH_LIMIT: 'Límite alto por hora (Contener)',
                 INPUT_HOURLY_HIGH_LIMIT: 'Por favor, introduzca el límite máximo por hora',
                 MASTER_METER: 'Medidor maestro',
@@ -10069,7 +10094,11 @@ function config($translateProvider) {
                 DATA_SOURCE_LIST: 'Список источников данных',
 
                 CONTROL_MODE_LIST: 'Control Mode List',
-                END_TIME_SHOULD_BE_AFTER_START_TIME: 'Время окончания должно быть после времени начала',
+
+				END_TIME_SHOULD_BE_AFTER_START_TIME: 'Время окончания должно быть после времени начала',
+				TARIFF_TIME_PERIODS_OVERLAP:'В один и тот же период времени разрешена только одна система расценок',
+				VALID_THROUGH_TIME_SHOULD_BE_AFTER_VALID_FROM_TIME: 'Дата окончания действия тарифа должна быть после даты начала',
+
                 IS_ACTIVE: 'Is Active',
                 START_TIME: 'Start Time',
                 END_TIME: 'End Time',
@@ -10925,6 +10954,7 @@ function config($translateProvider) {
             METER: {
                 HOURLY_LOW_LIMIT: 'Почасовой низкий лимит (Включительно)',
                 INPUT_HOURLY_LOW_LIMIT: 'Пожалуйста, введите почасовой низкий лимит',
+                INVALID_INPUT_HINT: 'Введите неправильные значения, диапазон значений от 0 до 999999',
                 HOURLY_HIGH_LIMIT: 'Почасовой максимальный лимит (Включительно)',
                 INPUT_HOURLY_HIGH_LIMIT: 'Пожалуйста, введите почасовой максимальный лимит',
                 MASTER_METER: 'Мастер-метр',
@@ -12020,7 +12050,11 @@ function config($translateProvider) {
                 DATA_SOURCE_LIST: 'قائمة مصادر البيانات',
 
                 CONTROL_MODE_LIST: 'Control Mode List',
-                END_TIME_SHOULD_BE_AFTER_START_TIME: 'يجب أن يكون وقت الانتهاء بعد وقت البدء',
+
+				END_TIME_SHOULD_BE_AFTER_START_TIME: 'يجب أن يكون وقت الانتهاء بعد وقت البدء',
+				TARIFF_TIME_PERIODS_OVERLAP:'يُسمح بنظام تسعير واحد فقط في نفس الفترة الزمنية',
+				VALID_THROUGH_TIME_SHOULD_BE_AFTER_VALID_FROM_TIME: 'يجب أن يكون تاريخ انتهاء سريان التعريفة بعد تاريخ البدء',
+
                 IS_ACTIVE: 'Is Active',
                 START_TIME: 'Start Time',
                 END_TIME: 'End Time',
@@ -12876,6 +12910,7 @@ function config($translateProvider) {
             METER: {
                 HOURLY_LOW_LIMIT: 'الحد الأدنى لكل ساعة (شامله)',
                 INPUT_HOURLY_LOW_LIMIT: 'يرجى إدخال الحد الأدنى للساعة',
+                INVALID_INPUT_HINT: 'القيمة المدخلة غير صحيحة ، مجموعة من 0 إلى 999999',
                 HOURLY_HIGH_LIMIT: 'الحد الأعلى للساعة (شامل)',
                 INPUT_HOURLY_HIGH_LIMIT: 'يرجى إدخال الحد الأعلى للساعة',
                 MASTER_METER: 'ماستر ميتر',
@@ -13971,7 +14006,11 @@ function config($translateProvider) {
                 DATA_SOURCE_LIST: 'Danh sách nguồn dữ liệu',
 
                 CONTROL_MODE_LIST: 'Control Mode List',
-                END_TIME_SHOULD_BE_AFTER_START_TIME: 'Thời gian kết thúc phải sau thời gian bắt đầu',
+
+				END_TIME_SHOULD_BE_AFTER_START_TIME: 'Thời gian kết thúc phải sau thời gian bắt đầu',
+				TARIFF_TIME_PERIODS_OVERLAP:'Chỉ cho phép một phương pháp tính giá trong cùng một khoảng thời gian',
+				VALID_THROUGH_TIME_SHOULD_BE_AFTER_VALID_FROM_TIME: 'Ngày kết thúc hiệu lực biểu phí phải sau ngày bắt đầu',
+
                 IS_ACTIVE: 'Is Active',
                 START_TIME: 'Start Time',
                 END_TIME: 'End Time',
@@ -14827,6 +14866,7 @@ function config($translateProvider) {
             METER: {
                 HOURLY_LOW_LIMIT: 'Giới hạn thấp hàng giờ (Bao gồm)',
                 INPUT_HOURLY_LOW_LIMIT: 'Vui lòng nhập giới hạn thấp hàng giờ',
+                INVALID_INPUT_HINT: 'Giá trị đầu vào không chính xác, phạm vi từ 0 đến 999999',
                 HOURLY_HIGH_LIMIT: 'Giới hạn cao hàng giờ (Bao gồm)',
                 INPUT_HOURLY_HIGH_LIMIT: 'Vui lòng nhập giới hạn cao hàng giờ',
                 MASTER_METER: 'Đồng hồ tổng thể',
@@ -15921,7 +15961,11 @@ function config($translateProvider) {
                 DATA_SOURCE_LIST: 'รายการแหล่งข้อมูล',
 
                 CONTROL_MODE_LIST: 'Control Mode List',
-                END_TIME_SHOULD_BE_AFTER_START_TIME: 'เวลาสิ้นสุดต้องอยู่หลังจากเวลาเริ่มต้น',
+
+				END_TIME_SHOULD_BE_AFTER_START_TIME: 'เวลาสิ้นสุดต้องอยู่หลังจากเวลาเริ่มต้น',
+				TARIFF_TIME_PERIODS_OVERLAP:'อนุญาตให้มีวิธีคำนวณราคาเพียงแบบเดียวในช่วงเวลาเดียวกัน',
+				VALID_THROUGH_TIME_SHOULD_BE_AFTER_VALID_FROM_TIME: 'วันที่สิ้นสุดอายุอัตราค่าใช้จ่ายต้องอยู่หลังจากวันที่เริ่มต้น',
+
                 IS_ACTIVE: 'Is Active',
                 START_TIME: 'Start Time',
                 END_TIME: 'End Time',
@@ -16776,6 +16820,7 @@ function config($translateProvider) {
             METER: {
                 HOURLY_LOW_LIMIT: 'ขีด จํากัด ต่ํารายชั่วโมง (รวม)',
                 INPUT_HOURLY_LOW_LIMIT: 'โปรดป้อนขีด จํากัด ต่ํารายชั่วโมง',
+                INVALID_INPUT_HINT: 'ตัวเลขที่ป้อนไม่ถูกต้องโดยมีค่าตั้งแต่ 0 ถึง 999,999',
                 HOURLY_HIGH_LIMIT: 'ขีด จํากัด สูงรายชั่วโมง (รวม)',
                 INPUT_HOURLY_HIGH_LIMIT: 'กรุณาใส่ขีด จํากัด สูงรายชั่วโมง',
                 MASTER_METER: 'มาสเตอร์มิเตอร์',
@@ -17871,7 +17916,11 @@ function config($translateProvider) {
                 DATA_SOURCE_LIST: 'Veri Kaynağı Listesi',
 
                 CONTROL_MODE_LIST: 'Control Mode List',
-                END_TIME_SHOULD_BE_AFTER_START_TIME: 'Bitiş saati başlangıç saatinden sonra olmalıdır',
+
+				END_TIME_SHOULD_BE_AFTER_START_TIME: 'Bitiş saati başlangıç saatinden sonra olmalıdır',
+				TARIFF_TIME_PERIODS_OVERLAP:'Aynı zaman diliminde yalnızca bir fiyatlandırma yöntemine izin verilir',
+				VALID_THROUGH_TIME_SHOULD_BE_AFTER_VALID_FROM_TIME: 'Tarife geçerlilik bitiş tarihi, başlangıç tarihinden sonra olmalıdır',
+
                 IS_ACTIVE: 'Is Active',
                 START_TIME: 'Start Time',
                 END_TIME: 'End Time',
@@ -18726,6 +18775,7 @@ function config($translateProvider) {
             METER: {
                 HOURLY_LOW_LIMIT: 'Saatlik Düşük Limit (Inclusive)',
                 INPUT_HOURLY_LOW_LIMIT: 'Lütfen saatlik düşük limiti girin',
+                INVALID_INPUT_HINT: 'Girilen değer 0 ile 999999 arasında yanlış.',
                 HOURLY_HIGH_LIMIT: 'Saatlik Yüksek Limit (Inclusive)',
                 INPUT_HOURLY_HIGH_LIMIT: 'Lütfen saatlik yüksek limiti girin',
                 MASTER_METER: 'Master Metre',
@@ -19821,7 +19871,11 @@ function config($translateProvider) {
                 DATA_SOURCE_LIST: 'Senarai Sumber Data',
 
                 CONTROL_MODE_LIST: 'Control Mode List',
-                END_TIME_SHOULD_BE_AFTER_START_TIME: 'Masa tamat mestilah selepas masa mula',
+
+				END_TIME_SHOULD_BE_AFTER_START_TIME: 'Masa tamat mestilah selepas masa mula',
+				TARIFF_TIME_PERIODS_OVERLAP:'Hanya satu kaedah penetapan harga dibenarkan dalam tempoh masa yang sama',
+				VALID_THROUGH_TIME_SHOULD_BE_AFTER_VALID_FROM_TIME: 'Tarikh tamat tempoh kadar mestilah selepas tarikh mula',
+
                 IS_ACTIVE: 'Is Active',
                 START_TIME: 'Start Time',
                 END_TIME: 'End Time',
@@ -20676,6 +20730,7 @@ function config($translateProvider) {
             METER: {
                 HOURLY_LOW_LIMIT: 'Had Rendah Setiap Jam (Termasuk)',
                 INPUT_HOURLY_LOW_LIMIT: 'Sila masukkan had rendah setiap jam',
+                INVALID_INPUT_HINT: 'Nilai yang dimasukkan tidak betul, dalam julat 0 hingga 999999',
                 HOURLY_HIGH_LIMIT: 'Had Tinggi Setiap Jam (Termasuk)',
                 INPUT_HOURLY_HIGH_LIMIT: 'Sila masukkan Had Tinggi Setiap Jam',
                 MASTER_METER: 'Master Meter',
@@ -21771,7 +21826,11 @@ function config($translateProvider) {
                 DATA_SOURCE_LIST: 'Daftar Sumber Data',
 
                 CONTROL_MODE_LIST: 'Control Mode List',
-                END_TIME_SHOULD_BE_AFTER_START_TIME: 'Waktu berakhir harus setelah waktu mulai',
+
+				END_TIME_SHOULD_BE_AFTER_START_TIME: 'Waktu berakhir harus setelah waktu mulai',
+				TARIFF_TIME_PERIODS_OVERLAP:'Hanya satu metode penetapan harga yang diizinkan dalam periode yang sama',
+				VALID_THROUGH_TIME_SHOULD_BE_AFTER_VALID_FROM_TIME: 'Tanggal berakhirnya tarif harus setelah tanggal mulai',
+
                 IS_ACTIVE: 'Is Active',
                 START_TIME: 'Start Time',
                 END_TIME: 'End Time',
@@ -22626,6 +22685,7 @@ function config($translateProvider) {
             METER: {
                 HOURLY_LOW_LIMIT: 'Batas Rendah Per Jam (Inclusive)',
                 INPUT_HOURLY_LOW_LIMIT: 'Harap masukkan Batas Rendah Per Jam',
+                INVALID_INPUT_HINT: 'Nilai yang dimasukkan tidak benar, berkisar dari 0 hingga 999999',
                 HOURLY_HIGH_LIMIT: 'Batas Tinggi Per Jam (Inclusive)',
                 INPUT_HOURLY_HIGH_LIMIT: 'Silakan masukkan batas tinggi per jam',
                 MASTER_METER: 'Pengukur Master',
@@ -23725,7 +23785,11 @@ function config($translateProvider) {
                 DATA_SOURCE_LIST: '資料來源清單',
 
                 CONTROL_MODE_LIST: 'Control Mode List',
-                END_TIME_SHOULD_BE_AFTER_START_TIME: '結束時間必須在開始時間之後',
+
+				END_TIME_SHOULD_BE_AFTER_START_TIME: '結束時間必須在開始時間之後',
+				TARIFF_TIME_PERIODS_OVERLAP:'同一時段只允許有一種計價方式',
+				VALID_THROUGH_TIME_SHOULD_BE_AFTER_VALID_FROM_TIME: '費率有效期的結束日期應在開始日期之後',
+
                 IS_ACTIVE: 'Is Active',
                 START_TIME: 'Start Time',
                 END_TIME: 'End Time',
@@ -24582,6 +24646,7 @@ function config($translateProvider) {
             METER: {
                 HOURLY_LOW_LIMIT: '每⼩時最⼩值（包含）',
                 INPUT_HOURLY_LOW_LIMIT: '請輸入每⼩時最⼩值',
+                INVALID_INPUT_HINT: '輸入的數值不正確，數值範圍為0到999999',
                 HOURLY_HIGH_LIMIT: '每⼩時最⼤值（包含）',
                 INPUT_HOURLY_HIGH_LIMIT: '請輸入每⼩時最⼤值',
                 MASTER_METER: '上級計量表',
@@ -25679,7 +25744,11 @@ function config($translateProvider) {
                 DATA_SOURCE_LIST: 'Lista de Fontes de Dados',
 
                 CONTROL_MODE_LIST: 'Control Mode List',
-                END_TIME_SHOULD_BE_AFTER_START_TIME: 'O horário de término deve ser após o horário de início',
+
+				END_TIME_SHOULD_BE_AFTER_START_TIME: 'O horário de término deve ser após o horário de início',
+				TARIFF_TIME_PERIODS_OVERLAP:'Apenas um método de precificação é permitido no mesmo período',
+				VALID_THROUGH_TIME_SHOULD_BE_AFTER_VALID_FROM_TIME: 'A data de término da vigência da tarifa deve ser posterior à data de início',
+
                 IS_ACTIVE: 'Is Active',
                 START_TIME: 'Start Time',
                 END_TIME: 'End Time',
@@ -26539,6 +26608,7 @@ function config($translateProvider) {
             METER: {
                 HOURLY_LOW_LIMIT: 'Limite Inferior Horário (Inclusivo)',
                 INPUT_HOURLY_LOW_LIMIT: 'Por favor, insira o Limite Inferior Horário',
+                INVALID_INPUT_HINT: 'O valor digitado é incorreto e varia de 0 a 999999',
                 HOURLY_HIGH_LIMIT: 'Limite Superior Horário (Inclusivo)',
                 INPUT_HOURLY_HIGH_LIMIT: 'Por favor, insira o Limite Superior Horário',
                 MASTER_METER: 'Medidor Mestre',