+++ myems-api/app.py
@@ -675,8 +675,8 @@ api.add_route('/photovoltaicpowerstations/{id_}/datasources',
               photovoltaicpowerstation.PhotovoltaicPowerStationDataSourceCollection())
 api.add_route('/photovoltaicpowerstations/{id_}/datasources/{dsid}',
               photovoltaicpowerstation.PhotovoltaicPowerStationDataSourceItem())
-
-
+api.add_route('/photovoltaicpowerstations/{id_}/datasourcepoints',
+              photovoltaicpowerstation.PhotovoltaicPowerStationDataSourcePointCollection())
 
 api.add_route('/points',
               point.PointCollection())
@@ -684,6 +684,8 @@ api.add_route('/points/{id_}',
               point.PointItem())
 api.add_route('/pointlimits/{id_}',
               point.PointLimit())
+api.add_route('/pointsetvalues/{id_}',
+              point.PointSetValue())
 api.add_route('/points/{id_}/export',
               point.PointExport())
 api.add_route('/points/import',