+++ myems-web/src/components/MyEMS/Tenant/TenantEnergyCategory.js
@@ -1321,6 +1321,38 @@ const TenantEnergyCategory = ({ setRedirect, setRedirectUrl, t }) => {
               <CountUp end={totalInTCO2E['value']} duration={2} prefix="" separator="," decimal="." decimals={2} />
             )}
           </CardSummary>
+          <CardSummary
+            rate={
+              totalInTCE['value'] && totalInTCE['value'] !== 0 && totalInTCO2E['value']
+                ? ((totalInTCO2E['value'] / totalInTCE['value'] - 1) * 100).toFixed(2) + '%'
+                : '--'
+            }
+            title={t('Reporting Period Consumption CATEGORY UNIT', {
+              CATEGORY: t('Carbon Emissions Per Unit Of Energy Consumption'),
+              UNIT: '(TCO2E/TCE)'
+            })}
+            color="warning"
+            footnote={t('Per Unit Area')}
+            footvalue={
+              totalInTCE['value_per_unit_area'] && totalInTCE['value_per_unit_area'] !== 0 && totalInTCO2E['value_per_unit_area']
+                ? (totalInTCO2E['value_per_unit_area'] / totalInTCE['value_per_unit_area']).toFixed(3)
+                : '--'
+            }
+            footunit="(TCO2E/TCE/M²)"
+          >
+            {totalInTCE['value'] && totalInTCE['value'] !== 0 && totalInTCO2E['value'] ? (
+              <CountUp
+                end={totalInTCO2E['value'] / totalInTCE['value']}
+                duration={2}
+                prefix=""
+                separator=","
+                decimal="."
+                decimals={3}
+              />
+            ) : (
+              '--'
+            )}
+          </CardSummary>
         </div>
         <Row noGutters>
           <Col className="mb-3 pr-lg-2 mb-3">